r4
forgot to wrap file deletion in a try-catch. now if it fails (for example, process is already running) it won't abort! hooray

I suppose I should try to prevent this error from occurring in the first place but it's not really important and I'm lazy so this'll have to do